+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2024-10-31
+
+## Changed
+
+- Fall back to SceneManager if Addressables load fails (editor only).
+
 ## [1.2.0] - 2024-05-14
 
 ## Added
@@ -19,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
-- Filter log messages by severity 
+- Filter log messages by severity
 
 ## Changed